refactor: centralize toast and dialog service in the root component

+++ src/app/components/expense/expense-card/expense-card.ts
@@ -3,20 +3,20 @@ import { CommonModule } from '@angular/common';
 import { BadgeModule } from 'primeng/badge';
 import { ButtonModule } from 'primeng/button';
 import { CardModule } from 'primeng/card';
-import { ConfirmDialogModule } from 'primeng/confirmdialog';
 import { Menu } from 'primeng/menu';
 import { TableModule } from 'primeng/table';
-import { ConfirmationService, MenuItem, MessageService } from 'primeng/api';
+import { MenuItem } from 'primeng/api';
 import { ExpenseViewType, ExpenseViewGroupType, isExpenseViewGroup } from '../../../domain/expense';
 import { CurrencyPipe } from '../../../pipes/currency-pipe';
 import { DatePipe } from '../../../pipes/date-pipe';
 import { ExpenseListing } from '../../../services/expense/expense-listing';
+import { Alert } from '../../../services/alert';
+import { ConfirmDialog } from '../../../services/confirm-dialog';
 
 @Component({
   selector: 'app-expense-card',
   imports: [
     CommonModule,
-    ConfirmDialogModule,
     BadgeModule,
     ButtonModule,
     CardModule,
@@ -27,14 +27,13 @@ import { ExpenseListing } from '../../../services/expense/expense-listing';
   ],
   templateUrl: './expense-card.html',
   styleUrl: './expense-card.css',
-  providers: [ConfirmationService],
 })
 export class ExpenseCard {
   @Input({ required: true, alias: 'value' }) expense!: ExpenseViewType | ExpenseViewGroupType;
 
-  private confirmSrv = inject(ConfirmationService);
+  private alert = inject(Alert);
+  private dialogSrv = inject(ConfirmDialog);
   private listing = inject(ExpenseListing);
-  private alert = inject(MessageService);
 
   protected collapse = signal(false);
 
@@ -47,7 +46,7 @@ export class ExpenseCard {
     {
       label: 'Supprimer',
       icon: 'pi pi-trash',
-      command: () => this.confirmDelete(),
+      command: () => this.deleteCommand(),
     },
   ];
 
@@ -59,44 +58,18 @@ export class ExpenseCard {
     throw 'TODO Edit method';
   }
 
-  private deleteExpense() {
-    this.listing
-      .remove(this.expense.id)
-      .then(() => {
-        this.alert.add({
-          severity: 'success',
-          summary: 'Succès',
-          detail: 'Une dépense a été supprimée.',
-        });
-      })
-      .catch(() => {
-        this.alert.add({
-          severity: 'error',
-          summary: 'Erreur',
-          detail: 'Erreur de suppression',
-        });
+  private deleteCommand() {
+    this.dialogSrv
+      .confirmDelete('Allez-vous supprimer cette dépense?')
+      .then((confirmed) => {
+        confirmed && this.deleteExpense();
       });
   }
 
-  private confirmDelete(): void {
-    this.confirmSrv.confirm({
-      position: 'top',
-      header: 'Confirmation',
-      message: 'Allez-vous supprimer cette dépense?',
-      icon: 'pi pi-exclamation-triangle',
-      rejectLabel: 'Annuler',
-      rejectButtonProps: {
-        label: 'NON',
-        severity: 'secondary',
-        outlined: true,
-      },
-      acceptButtonProps: {
-        label: 'OUI',
-        severity: 'danger',
-      },
-      accept: () => {
-        this.deleteExpense();
-      },
-    });
+  private deleteExpense() {
+    this.listing
+      .remove(this.expense.id)
+      .then(() => this.alert.success({ detail: 'Une dépense a été supprimée.' }))
+      .catch(() => this.alert.error({ detail: 'Erreur de suppression' }));
   }
 }
